[network] Address Koan review: platform IPv4 gate, nRF52 IPv6 invariant, misc fixes
The automatic IPv4 drop had no platform/framework gate, so it silently fired
on host, esp8266, rp2, bk72xx/ln882x/rtl87xx, and esp32-arduino -- none of
which have any build-system wiring to actually strip IPv4, so nothing was
saved and socket.cpp's IPv4 handling broke for no benefit. Fixed by having
each platform declare its own requirement via network.require_ipv4 in its
own CONFIG_SCHEMA (esp8266, rp2, host directly; libretiny's shared
BASE_SCHEMA covers bk72xx/ln882x/rtl87xx at once; esp32 only when using the
Arduino framework, since CONFIG_LWIP_IPV4 has no Arduino override the way
CONFIG_LWIP_IPV6 does) instead of hardcoding a platform gate inside
'network' itself -- same architecture as the component-level registration,
one level up. network/__init__.py's to_code() needed no changes: the
existing has_ipv4_requirement() formula was already correct once callers
exist.

Same fix for nRF52's IPv6 invariant, deleted along with the old
SplitDefault(nrf52=True) + validate_ipv4 without a replacement: nrf52 now
calls network.require_ipv6 unconditionally in its own CONFIG_SCHEMA, so any
nRF52 config (not just ones using openthread) defaults to IPv6 on / IPv4 off
again.

Also from the same review:
- wake_on_lan (button platform) unconditionally broadcasts over IPv4 but
  never registered a requirement -- registered it.
- _detect_explicit_ipv6_disable only matched the literal Python `False`, so
  a substitution or quoted 'false'/'no'/'off' silently bypassed the
  explicit-disable-vs-requirement conflict check. Normalized with
  cv.boolean() first.
- Fixed signed-shift UB in the two new Zephyr 4-octet IPAddress
  constructors (uint8_t operand promotes to signed int before <<, undefined
  for octets >= 128) by casting to uint32_t first, matching the existing
  host constructor.

+++ esphome/components/network/ip_address.h
@@ -100,7 +100,7 @@ struct IPAddress {
 #if defined(CONFIG_NET_IPV4)
   IPAddress(uint8_t first, uint8_t second, uint8_t third, uint8_t fourth) {
     struct in_addr addr4 {
-      .s_addr = htonl((first << 24) | (second << 16) | (third << 8) | fourth)
+      .s_addr = htonl(((uint32_t) first << 24) | ((uint32_t) second << 16) | ((uint32_t) third << 8) | fourth)
     };
     net_ipv6_addr_create_v4_mapped(&addr4, &ip_addr_);
   }
@@ -144,7 +144,8 @@ struct IPAddress {
   // uses for the dual-stack (V4-mapped) and IPv6-only cases.
   IPAddress() { ip_addr_.s_addr = 0; }
   IPAddress(uint8_t first, uint8_t second, uint8_t third, uint8_t fourth) {
-    this->ip_addr_.s_addr = htonl((first << 24) | (second << 16) | (third << 8) | fourth);
+    this->ip_addr_.s_addr =
+        htonl(((uint32_t) first << 24) | ((uint32_t) second << 16) | ((uint32_t) third << 8) | fourth);
   }
   IPAddress(const std::string &in_address) { ipaddr_aton(in_address.c_str(), &ip_addr_); }
   IPAddress(const ip_addr_t *other_ip) { ip_addr_ = *other_ip; }
